Add basic 'album' support
An Album is a model that represents a collection of uploaded files.
Currently the only way the user can create albums is by uploading
multiple files. This will create an album consisting of each of those
uploaded files, which can be accessed with a single request and viewed
exactly like the gallery or viewmine views. There, I did it. Just like
you asked.